CSW-to-Kenna/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
importlib-resources==5.4.0
Jinja2==3.0.3
MarkupSafe==2.0.1
netaddr==0.10.1
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
resolvelib==0.5.4
six==1.16.0
tetpyclient==1.0.8
tqdm==4.64.0
urllib3==1.26.9
zipp==3.6.0
